requirements.txt
torch>=2.4.0
numpy>=1.26.0
scipy>=1.13.0
scikit-learn>=1.5.0
pandas>=2.2.0
matplotlib>=3.9.0
tqdm>=4.66.0
pyyaml>=6.0.0
normflows>=1.7.0